pyproject.toml
[project]
name = "agent-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dependency-injector>=4.48.3",
    "fastapi>=0.127.0",
    "gtts>=2.5.4",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "rich>=14.2.0",
    "twilio>=9.9.0",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
